This adds a new variation card component and moves the product variation types form to it, so we can start implementing other parts of the variations UI and keep them contained.
